added new route for config vars
